requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-extensions==2.1.4
djangorestframework==3.9.0
firebase-admin==2.16.0
idna==2.7
kombu==4.2.1
numpy==1.15.4
pytz==2018.7
redis==3.0.1
requests==2.20.1
six==1.11.0
urllib3==1.24.1
vine==1.1.4
